Fix build error showing up on Travis

+++ src/ExtraRegisters.h
@@ -56,7 +56,7 @@ public:
     arch_ = a;
     format_ = format;
     std::swap(data_, consume_data);
-    static const int min_xsave_size = 576;
+    static const size_t min_xsave_size = 576;
     assert(format == NONE || data_.size() >= min_xsave_size);
   }
